requirements.txt
et_xmlfile==2.0.0
numpy==2.0.2
openpyxl==3.1.5
pandas==2.3.2
python-dateutil==2.9.0.post0
pytz==2025.2
scipy==1.13.1
six==1.17.0
tzdata==2025.2
PyYAML==6.0.2
